refactor: reorganize domain entities into dedicated packages
