--- backend/src/main/java/br/com/calendar/user/dto/UserResponse.java
@@ -1,4 +1,0 @@
-package br.com.calendar.user.dto;
-
-public record UserResponse(String id, String name, String email) {
-}
