+++ backend/src/main/java/br/com/calendar/user/dto/UserSummaryDTO.java
@@ -1,0 +1,4 @@
+package br.com.calendar.user.dto;
+
+public record UserSummaryDTO(String id, String name, String email) {
+}
